backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
python-multipart==0.0.9
pytest==8.2.0
httpx==0.27.0
transformers==4.44.2
torch==2.4.1
sacremoses
pypdf>=3.0.0
python-docx>=1.1.0